requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Data Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0

# Security & Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.1.3

# HTTP & File Upload
python-multipart>=0.0.20
httpx>=0.28.0

# Database
pymongo>=4.10.0
motor>=3.6.0

# Environment Variables
python-dotenv>=1.0.1

# Email (optional - simplified alternative)
# emails>=0.6

# Google OAuth
google-auth>=2.37.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0